[development/imgui_3] renaming imgui_render2 to imgui_renderer
